Backend/pyproject.toml
[project]
name = "ragvoiceagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.4",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pipecat-ai[deepgram,openai,elevenlabs,silero]>=0.0.60",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "pypdf>=6.7.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
